@@ -79,30 +57,32 @@
   },
   "homepage": "https://github.com/commercetools/commercetools-sunrise-data#readme",
   "devDependencies": {
-    "@commercetools/api-request-builder": "5.0.0",
-    "@commercetools/sdk-client": "2.0.0",
-    "@commercetools/sdk-middleware-auth": "5.0.0",
-    "@commercetools/sdk-middleware-http": "5.0.0",
-    "babel-cli": "6.26.0",
-    "babel-preset-env": "1.7.0",
-    "cross-env": "^5.2.0",
+    "@babel/core": "^7.7.5",
+    "@babel/plugin-proposal-object-rest-spread": "^7.7.4",
+    "@commercetools/api-request-builder": "5.3.2",
+    "@commercetools/sdk-client": "2.1.0",
+    "@commercetools/sdk-middleware-auth": "5.1.2",
+    "@commercetools/sdk-middleware-http": "5.2.1",
+    "@babel/cli": "^7.7.5",
+    "@babel/preset-env": "^7.7.6",
+    "braces": "^3.0.2",
+    "dotenv": "^8.0.0",
+    "fstream": "^1.0.12",
     "nconf": "0.10.0",
     "path": "0.12.7",
     "rimraf": "2.6.3",
     "snazzy": "8.0.0",
-    "sphere-node-cli": "^2.1.7",
-    "sphere-node-product-csv-sync": "1.5.4",
-    "sphere-stock-import": "^2.1.0",
-    "standard": "12.0.1"
+    "standard": "13.1.0"
   },
   "dependencies": {
-    "@commercetools/sdk-auth": "^2.0.0",
-    "csvtojson": "^2.0.8",
-    "es6-promise": "^4.2.5",
+    "@commercetools/sdk-auth": "^2.1.2",
+    "cli-progress": "^3.4.0",
+    "csvtojson": "^2.0.10",
+    "es6-promise": "^4.2.8",
     "isomorphic-fetch": "^2.2.1",
-    "sphere-category-sync": "^1.2.0"
+    "core-js": "3"
   },
   "engines": {
-    "node": "^8"
+    "node": "^10"
   }
 }
