@@ -1,11 +1,6 @@
-import {
-  client,
-  customersService,
-  customerGroupService
-} from './services'
-var nconf = require('nconf')
-
-var customersData = require(process.cwd() + '/data/customers.json')
+import { client, customersService, customerGroupService } from './services'
+const nconf = require('nconf')
+const customersData = require(process.cwd() + '/data/customers.json')
 
 const addCustomer = (cData) => {
   const updateRequest = {
@@ -32,7 +27,7 @@ const importCustomerGroups = () => {
   })
 }
 
-const deleteAllCustomers = () => {
+const deleteAllCustomers = (cb) => {
   const request = {
     uri: customersService
       .build(),
@@ -56,11 +51,14 @@ const deleteAllCustomers = () => {
     }
   ).then(() => {
     console.log('\x1b[32m%s\x1b[0m', 'All customers are succesfully deleted')
+    if (cb) {
+      return cb()
+    }
   })
     .catch(err => console.log(err))
 }
 
-const deleteCustomerGroups = () => {
+const deleteCustomerGroups = (cb) => {
   const request = {
     uri: customerGroupService
       .build(),
@@ -84,6 +82,7 @@ const deleteCustomerGroups = () => {
     }
   ).then(() => {
     console.log('\x1b[32m%s\x1b[0m', 'All customer groups are succesfully deleted')
+    return cb()
   })
     .catch(err => console.log(err))
 }
@@ -92,8 +91,14 @@ if (nconf.get('clean')) {
   deleteAllCustomers()
   deleteCustomerGroups()
 } else if (nconf.get('import')) {
+  console.log('\x1b[32m%s\x1b[0m', 'Importing customers...')
   importCustomerGroups()
   customersData.forEach(element => {
     addCustomer(element)
   })
+}
+
+module.exports = {
+  deleteAllCustomers,
+  deleteCustomerGroups
 }
