@@ -1,15 +1,11 @@
-import {
-  client,
-  typesService
-} from './services'
-
-var nconf = require('nconf')
+import { client, typesService } from './services'
+const nconf = require('nconf')
 
 const importTypes = () => {
-  var channelTypes = require(process.cwd() + '/data/channel-types.json')
-  var customerTypes = require(process.cwd() + '/data/customer-types.json')
-  var orderTypes = require(process.cwd() + '/data/order-types.json')
-  var allTypes = [channelTypes, customerTypes, orderTypes]
+  const channelTypes = require(process.cwd() + '/data/channel-types.json')
+  const customerTypes = require(process.cwd() + '/data/customer-types.json')
+  const orderTypes = require(process.cwd() + '/data/order-types.json')
+  const allTypes = [channelTypes, customerTypes, orderTypes]
 
   allTypes.forEach(type => {
     type.forEach(element => {
@@ -24,7 +20,7 @@ const importTypes = () => {
   })
 }
 
-const deleteTypes = () => {
+const deleteTypes = (cb) => {
   const request = {
     uri: typesService
       .build(),
@@ -48,6 +44,9 @@ const deleteTypes = () => {
     }
   ).then(() => {
     console.log('\x1b[32m%s\x1b[0m', 'All types are succesfully deleted')
+    if (cb) {
+      return cb()
+    }
   })
     .catch(err => console.log(err))
 }
@@ -55,5 +54,9 @@ const deleteTypes = () => {
 if (nconf.get('clean')) {
   deleteTypes()
 } else if (nconf.get('import')) {
+  console.log('\x1b[32m%s\x1b[0m', 'Importing types...')
   importTypes()
+}
+module.exports = {
+  deleteTypes
 }
