@@ -1,9 +1,7 @@
-import {
-  client,
-  cartsService
-} from './services'
+import { client, cartsService } from './services'
+const nconf = require('nconf')
 
-const deleteAllCarts = () => {
+const deleteAllCarts = (cb) => {
   const request = {
     uri: cartsService
       .build(),
@@ -27,8 +25,17 @@ const deleteAllCarts = () => {
     }
   ).then(() => {
     console.log('\x1b[32m%s\x1b[0m', 'All carts are succesfully deleted')
+    if (cb) {
+      return cb()
+    }
   })
     .catch(err => console.log(err))
 }
 
-deleteAllCarts()
+if (nconf.get('clean')) {
+  deleteAllCarts()
+}
+
+module.exports = {
+  deleteAllCarts
+}
