@@ -1,7 +1,7 @@
 import { client, channelsService } from './services'
-var nconf = require('nconf')
+const nconf = require('nconf')
 
-const deleteChannels = () => {
+const deleteChannels = (cb) => {
   const request = {
     uri: channelsService
       .build(),
@@ -25,6 +25,9 @@ const deleteChannels = () => {
     }
   ).then(() => {
     console.log('\x1b[32m%s\x1b[0m', 'All channels are succesfully deleted')
+    if (cb) {
+      return cb()
+    }
   })
     .catch(err => console.log(err))
 }
@@ -46,5 +49,10 @@ const importChannels = () => {
 if (nconf.get('clean')) {
   deleteChannels()
 } else if (nconf.get('import')) {
+  console.log('\x1b[32m%s\x1b[0m', 'Importing channels...')
   importChannels()
+}
+
+module.exports = {
+  deleteChannels
 }
