@@ -1,4 +1,5 @@
 var nconf = require('nconf')
+require('dotenv').config()
 var {
   URL
 } = require('url')
@@ -28,13 +29,13 @@ class Configuration {
       .env()
 
     // required key check
-    nconf.required(['projectKey', 'clientId', 'clientSecret', 'authUrl', 'apiUrl'])
+    // nconf.required(['projectKey', 'clientId', 'clientSecret', 'authUrl', 'apiUrl'])
     this.config = {
-      projectKey: nconf.get('projectKey'),
-      clientId: nconf.get('clientId'),
-      clientSecret: nconf.get('clientSecret'),
-      authUrl: this.normalizeUrl(nconf.get('authUrl')),
-      apiUrl: this.normalizeUrl(nconf.get('apiUrl'))
+      projectKey: process.env.CTP_PROJECT_KEY,
+      clientId: process.env.CTP_CLIENT_ID,
+      clientSecret: process.env.CTP_CLIENT_SECRET,
+      authUrl: this.normalizeUrl(process.env.CTP_AUTH_URL),
+      apiUrl: this.normalizeUrl(process.env.CTP_API_URL)
     }
   }
 
