@@ -1,9 +1,10 @@
-import {
-  client,
-  inventoryService
-} from './services'
+import { client, inventoryService } from './services'
+require('dotenv').config()
+const nconf = require('nconf')
+const execSync = require('child_process').execSync
+const path = require('path')
 
-const deleteInventory = () => {
+const deleteInventory = (cb) => {
   const request = {
     uri: inventoryService
       .build(),
@@ -27,8 +28,26 @@ const deleteInventory = () => {
     }
   ).then(() => {
     console.log('\x1b[32m%s\x1b[0m', 'All inventory entries are succesfully deleted')
+    if (cb) {
+      return cb()
+    }
   })
     .catch(err => console.log(err))
 }
 
-deleteInventory()
+const importInventory = (cmd) => {
+  console.log('Importing inventory...')
+  return path.join('node_modules', '.bin', cmd)
+}
+
+if (nconf.get('clean')) {
+  deleteInventory()
+} else if (nconf.get('import')) {
+  console.log('\x1b[32m%s\x1b[0m', 'Importing inventory...')
+  execSync(`${importInventory('stock-import')} --logLevel debug --logSilent true --file data/inventory.csv --projectKey ${process.env.CTP_PROJECT_KEY} --clientId ${process.env.CTP_CLIENT_ID} --clientSecret ${process.env.CTP_CLIENT_SECRET} --sphereHost https://${process.env.CTP_API_URL} --sphereAuthHost https://${process.env.CTP_AUTH_URL}`)
+  execSync(`${importInventory('stock-import')} --logLevel debug --logSilent true --file data/inventory-stores.csv --projectKey ${process.env.CTP_PROJECT_KEY} --clientId ${process.env.CTP_CLIENT_ID} --clientSecret ${process.env.CTP_CLIENT_SECRET} --sphereHost https://${process.env.CTP_API_URL} --sphereAuthHost https://${process.env.CTP_AUTH_URL}`)
+}
+
+module.exports = {
+  deleteInventory
+}
