@@ -1,7 +1,8 @@
 import { client, stateService } from './services'
-var nconf = require('nconf')
+const nconf = require('nconf')
 
-var initialId
+let initialId
+
 const loadInitialStateId = () => {
   client.execute({
     uri: stateService.where('builtIn = true').build(),
@@ -10,41 +11,46 @@ const loadInitialStateId = () => {
     .catch(error => console.log(error))
 }
 
-const clearAllStatesWithTransaction = () => {
+const clearAllStatesWithTransaction = (cb) => {
   client.execute({
     uri: stateService.where('transitions is not empty').build(),
     method: 'GET'
-  }).then(result => unsetTransitions(result.body.results))
+  }).then(result => unsetTransitions(result.body.results, cb))
     .catch(error => console.log(error))
 }
 
-const unsetTransitions = (results) => {
-  var promises = []
+const unsetTransitions = (results, cb) => {
+  const promises = []
   const action = [{
+    /* eslint quote-props: ["error", "consistent"] */
     'action': 'setTransitions',
     'transitions': []
   }]
-  for (var i = 0, len = results.length; i < len; i++) {
+  for (let i = 0, len = results.length; i < len; i++) {
     promises.push(client.execute({
       uri: stateService.byId(results[i].id).withVersion(results[i].version).build(),
       method: 'POST',
       body: { version: results[i].version, actions: action }
     }))
   }
-  // console.log('promises.length == ' + promises.length)
   Promise.all(promises)
     .then(res => {
       client.execute({
         uri: stateService.where('builtIn = false').build(),
         method: 'GET'
       }).then(response => {
-        var deletePromises = []
-        for (var i = 0, len = response.body.results.length; i < len; i++) {
+        const deletePromises = []
+        for (let i = 0, len = response.body.results.length; i < len; i++) {
           deletePromises.push(deleteStates(response.body.results[i]))
         }
 
         Promise.all(deletePromises)
-          .then(res => console.log('\x1b[32m%s\x1b[0m', 'All states are succesfully deleted'))
+          .then(res => {
+            console.log('\x1b[32m%s\x1b[0m', 'All states are succesfully deleted')
+            if (cb) {
+              return cb()
+            }
+          })
           .catch(err => console.log(err))
       })
     })
@@ -69,7 +75,7 @@ const createStateDraft = (key, nameEN, nameDE) => {
 }
 
 const createTransactionsDraft = (toIds) => {
-  var transactionObj = []
+  const transactionObj = []
   toIds.forEach(item => {
     transactionObj.push({ 'typeId': 'state', 'id': item })
   })
@@ -107,7 +113,7 @@ const createState = (key, nameEN, nameDE) => {
 }
 
 const AddStateToProject = () => {
-  var readyToShipId, backOrderId, shippedId, cancelledId, pickingId, returnedId, returnApprovedId,
+  let readyToShipId, backOrderId, shippedId, cancelledId, pickingId, returnedId, returnApprovedId,
     returnNotApprovedId, closedId, lostId, lossApprovedId, lossNotApprovedId
 
   Promise.all([createState('readyToShip', 'Ready to Ship', 'Versandfertig'),
@@ -153,6 +159,11 @@ const AddStateToProject = () => {
 if (nconf.get('clean')) {
   clearAllStatesWithTransaction()
 } else if (nconf.get('import')) {
+  console.log('\x1b[32m%s\x1b[0m', 'Importing line item states...')
   loadInitialStateId()
   AddStateToProject()
+}
+
+module.exports = {
+  clearAllStatesWithTransaction
 }
