@@ -1,12 +1,7 @@
-import {
-  client,
-  ordersService,
-  ordersImportService
-} from './services'
+import { client, ordersService, ordersImportService } from './services'
+const nconf = require('nconf')
 
-var nconf = require('nconf')
-
-const deleteAllOrders = () => {
+const deleteAllOrders = (cb) => {
   const request = {
     uri: ordersService
       .build(),
@@ -30,6 +25,9 @@ const deleteAllOrders = () => {
     }
   ).then(() => {
     console.log('\x1b[32m%s\x1b[0m', 'All orders are succesfully deleted')
+    if (cb) {
+      return cb()
+    }
   })
     .catch(err => console.log(err))
 }
@@ -55,6 +53,7 @@ const lineItemDraft = (item) => {
 
 const orderDraft = (item) => {
   return {
+    /* eslint quote-props: ["error", "consistent"] */
     'customerEmail': item.customerEmail,
     'orderNumber': item.orderNumber,
     'lineItems': [{
@@ -106,7 +105,7 @@ const importOrders = (csvFilePath) => {
           body: item
         }
         client.execute(request)
-          .catch(err => console.log(err.body.errors))
+          .catch(err => console.log(err))
       })
     })
 }
@@ -114,5 +113,10 @@ const importOrders = (csvFilePath) => {
 if (nconf.get('clean')) {
   deleteAllOrders()
 } else if (nconf.get('import')) {
+  console.log('\x1b[32m%s\x1b[0m', 'Importing orders...')
   importOrders(nconf.get('csv'))
+}
+
+module.exports = {
+  deleteAllOrders
 }
