@@ -1,7 +1,6 @@
 import { client, shippingMethodService, taxService, zonesService } from './services'
-var nconf = require('nconf')
-
-var shippingMethods = require(process.cwd() + '/data/shipping-methods.json')
+const nconf = require('nconf')
+const shippingMethods = require(process.cwd() + '/data/shipping-methods.json')
 
 const getTaxCategories = () => {
   return client.execute({
@@ -35,7 +34,7 @@ const addShippingMethods = (taxCategoriesByName, zonesForName) => {
   })
 }
 
-const deleteAllShippingMethods = () => {
+const deleteAllShippingMethods = (cb) => {
   const request = {
     uri: shippingMethodService
       .build(),
@@ -59,19 +58,22 @@ const deleteAllShippingMethods = () => {
     }
   ).then(() => {
     console.log('\x1b[32m%s\x1b[0m', 'All shipping methods are succesfully deleted')
+    if (cb) {
+      return cb()
+    }
   })
     .catch(err => console.log(err))
 }
 
 async function run () {
   try {
-    let taxCategories = await getTaxCategories()
+    const taxCategories = await getTaxCategories()
     const taxCategoriesByName = taxCategories.body.results.reduce((obj, taxCategory) => {
       obj[taxCategory.name] = taxCategory
       return obj
     }, {})
 
-    let zones = await getZones()
+    const zones = await getZones()
     const zonesForName = zones.body.results.reduce((obj, zone) => {
       obj[zone.name] = zone
       return obj
@@ -82,10 +84,15 @@ async function run () {
     console.log(e)
     throw e // let caller know the promise rejected with this reason
   }
-}
+};
 
 if (nconf.get('clean')) {
   deleteAllShippingMethods()
 } else if (nconf.get('import')) {
+  console.log('\x1b[32m%s\x1b[0m', 'Importing shipping methods...')
   run()
+}
+
+module.exports = {
+  deleteAllShippingMethods
 }
