@@ -1,7 +1,7 @@
 import { client, taxService } from './services'
-var nconf = require('nconf')
+const nconf = require('nconf')
 
-const deleteTaxCategories = () => {
+const deleteTaxCategories = (cb) => {
   const request = {
     uri: taxService
       .build(),
@@ -25,6 +25,9 @@ const deleteTaxCategories = () => {
     }
   ).then(() => {
     console.log('\x1b[32m%s\x1b[0m', 'All tax categories are succesfully deleted')
+    if (cb) {
+      return cb()
+    }
   })
     .catch(err => console.log(err))
 }
@@ -46,5 +49,10 @@ const importTaxCategories = () => {
 if (nconf.get('clean')) {
   deleteTaxCategories()
 } else if (nconf.get('import')) {
+  console.log('\x1b[32m%s\x1b[0m', 'Importing tax categories...')
   importTaxCategories()
+}
+
+module.exports = {
+  deleteTaxCategories
 }
