@@ -1,7 +1,6 @@
 import { client, zonesService } from './services'
-var nconf = require('nconf')
-
-var zones = require(process.cwd() + '/data/zones.json')
+const nconf = require('nconf')
+const zones = require(process.cwd() + '/data/zones.json')
 
 const addZones = () => {
   zones.forEach(element => {
@@ -16,7 +15,7 @@ const addZones = () => {
   })
 }
 
-const deleteAllZones = () => {
+const deleteAllZones = (cb) => {
   const request = {
     uri: zonesService
       .build(),
@@ -40,6 +39,9 @@ const deleteAllZones = () => {
     }
   ).then(() => {
     console.log('\x1b[32m%s\x1b[0m', 'All zones are succesfully deleted')
+    if (cb) {
+      return cb()
+    }
   })
     .catch(err => console.log(err))
 }
@@ -47,5 +49,10 @@ const deleteAllZones = () => {
 if (nconf.get('clean')) {
   deleteAllZones()
 } else if (nconf.get('import')) {
+  console.log('\x1b[32m%s\x1b[0m', 'Importing zones...')
   addZones()
+}
+
+module.exports = {
+  deleteAllZones
 }
